Revise the Corchuelo et al. description.
Previously I found the algorithm hard to read because splitting it over two
columns doesn't work well with Python's indentation syntax. So instead of
splitting the algorithm over two columns, in this commit I put the algorithm in
one column and the caption in the other column. I can't work out whether this is
a good idea or not, but I think it's worth living with it for a few days and
then deciding.