The description of duplicate detection was wrong.
In fact, I suspect the original implementation was probably wrong, as it didn't
take into account N_shifts. However, we long since ditched that implementation
in favour of compatible node merging. I'd left the description in the paper
because I thought it helped warm readers up to compatible node merging. However,
fixing the description of duplicate detection made it almost as fiddly as
compatible node merging. So... off with duplicate detection's head!