Import new draft data.
This is from a loaded machine, with turbo boost on, and only running 3 pexecs --
but experience suggests these numbers will be fairly representative of those
we'll get from the dedicated benchmarking machines.

The major eye-opener here is, perhaps unsurprisingly, from the Corchuelo
recoverer, which doesn't fit the expected pattern with tokens skipped. That
means we have to revise how we view Hypothesis H3.